requirements.txt
# Async File Management
aiofiles==23.2.1

# Async HTTP Requests
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
frozenlist==1.4.1
httpcore==1.0.5
httpx==0.27.0
sniffio==1.3.1

# Async Cache
aioredis==1.3.1
redis==5.0.6

# Database Management
alembic==1.13.1
asyncpg==0.29.0
pgcli==4.1.0
pgspecial==2.1.2
psycopg==3.1.19
SQLAlchemy==2.0.30
sqlparse==0.5.0
sqladmin==0.17.0

# Celery and Messaging
amqp==5.2.0
billiard==4.2.0
celery==5.4.0
flower==2.0.1
kombu==5.3.7
vine==5.1.0

# API Frameworks and Utilities
fastapi==0.111.0
fastapi-cache==0.1.0
fastapi-cache2==0.2.1
fastapi-cli==0.0.4
fastapi-versioning==0.10.0
starlette==0.37.2

# Web Servers and ASGI
gunicorn==22.0.0
httptools==0.6.1
uvicorn==0.29.0
uvloop==0.19.0
websockets==12.0

# Testing and Linting
autoflake==2.3.1
black==24.4.2
flake8==7.1.0
isort==5.13.2
mypy-extensions==1.0.0
pycodestyle==2.12.0
pyflakes==3.2.0
pyright==1.1.370
pytest==8.2.2
pytest-asyncio==0.23.7
time-machine==2.14.1

# Security and Authentication
bcrypt==4.1.3
ecdsa==0.19.0
email_validator==2.1.1
passlib==1.7.4
python-jose==3.3.0
rsa==4.9

# Development Tools
appdirs==1.4.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
jedi==0.19.1
nodeenv==1.9.1
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
pydantic==2.7.2
pydantic-settings==2.2.1
pydantic_core==2.18.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rich==13.7.1
shellingham==1.5.4
tabulate==0.9.0
typer==0.12.3
typing_extensions==4.12.0
watchfiles==0.22.0

# Logging and Monitoring
prometheus_client==0.20.0
python-json-logger==2.0.7
sentry-sdk==2.7.1
prometheus_fastapi_instrumentator==7.0.0

# Templates and HTML
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
WTForms==3.1.2

# Markdown and Documentation
markdown-it-py==3.0.0
mdurl==0.1.2

# CLI Tools
cli_helpers==2.3.1
ptpython==3.0.27
prompt_toolkit==3.0.45
wcwidth==0.2.13

# Miscellaneous (with Explanations)
attrs==23.2.0          # Data validation and serialization
certifi==2024.2.2      # SSL certificates
configobj==5.0.8       # Configuration files
dnspython==2.6.1       # DNS operations
greenlet==3.0.3        # Lightweight concurrency
h11==0.14.0            # HTTP/1.1 protocol
hiredis==2.3.2         # Redis client
humanize==4.9.0        # Human-readable formatting
iniconfig==2.0.0       # .ini files parsing
itsdangerous==2.2.0    # Data integrity
orjson==3.10.3         # JSON serialization
parso==0.8.4           # Python code parsing
pathspec==0.12.1       # File path pattern matching
pillow==10.3.0         # Image processing
PyYAML==6.0.1          # YAML parsing
tornado==6.4.1         # Web framework and async networking
tzdata==2024.1         # Time zone data
urllib3==2.2.2         # HTTP library
yarl==1.9.4            # URL parsing
setproctitle==1.3.3    # Customize process title
mccabe==0.7.0          # flake8 plugin
